Replaced f64 with f32 -> +32% speed on rectangles benchmark

+++ ray_tracer/src/aabb.rs
@@ -28,7 +28,7 @@ impl AABB {
     /// check for collisions, using the slab method
     pub fn collide(&self, ray: Ray) -> bool {
         let mut t_min = 0.0;
-        let mut t_max = f64::MAX;
+        let mut t_max = f32::MAX;
 
         for i in 0..3 as usize {
             /*
@@ -37,7 +37,7 @@ impl AABB {
             If the ray is inside the interval, then it could intersect on
             a different axis.
             */
-            if ray.direction[i] == 0f64 {
+            if ray.direction[i] == 0f32 {
                 if ray.origin[i] < self.minimum[i] || self.maximum[i] <= ray.origin[i] {
                     return false;
                 }
@@ -48,10 +48,10 @@ impl AABB {
             let mut t0 = (self.minimum[i] - ray.origin[i]) / ray.direction[i];
             let mut t1 = (self.maximum[i] - ray.origin[i]) / ray.direction[i];
 
-            (t0, t1) = (f64::min(t0, t1), f64::max(t0, t1));
+            (t0, t1) = (f32::min(t0, t1), f32::max(t0, t1));
 
-            t_min = f64::max(t_min, t0);
-            t_max = f64::min(t_max, t1);
+            t_min = f32::max(t_min, t0);
+            t_max = f32::min(t_max, t1);
 
             if t_min > t_max {
                 return false;
@@ -64,9 +64,9 @@ impl AABB {
 
     /*
     /// check for collisions, using the slab method
-      fn hit(&self, ray: &Ray, mut t_min: f64, mut t_max: f64) -> bool {
+      fn hit(&self, ray: &Ray, mut t_min: f32, mut t_max: f32) -> bool {
           for i in 0..3 as usize {
-              let (t0, t1) = if ray.direction()[i] == 0f64 {
+              let (t0, t1) = if ray.direction()[i] == 0f32 {
                   /*
                   The ray does not change its position along the i axis. Therefore,
                   the origin of the ray determines whether the ray is going to be.
@@ -82,11 +82,11 @@ impl AABB {
                   let t0 = (self.minimum[i] - ray.origin()[i]) / ray.direction()[i];
                   let t1 = (self.maximum[i] - ray.origin()[i]) / ray.direction()[i];
 
-                  (f64::min(t0, t1), f64::max(t0, t1))
+                  (f32::min(t0, t1), f32::max(t0, t1))
               };
 
-              t_min = f64::max(t_min, t0);
-              t_max = f64::min(t_max, t1);
+              t_min = f32::max(t_min, t0);
+              t_max = f32::min(t_max, t1);
 
               if t_min >= t_max {
                   return false;
