+++ ray_tracer/src/image.rs
@@ -96,18 +96,18 @@ impl From<Color> for ColorU8 {
 
 #[derive(Copy, Clone, Serialize, Deserialize)]
 pub struct Color {
-    pub red: f64,
-    pub green: f64,
-    pub blue: f64,
+    pub red: f32,
+    pub green: f32,
+    pub blue: f32,
 }
 
 impl Color {
-    pub const fn new(red: f64, green: f64, blue: f64) -> Color {
+    pub const fn new(red: f32, green: f32, blue: f32) -> Color {
         Color { red, green, blue }
     }
 
-    pub fn to_u8(channel: f64) -> u8 {
-        u8::min(255, (channel * 255f64) as u8)
+    pub fn to_u8(channel: f32) -> u8 {
+        u8::min(255, (channel * 255f32) as u8)
     }
 
     pub fn gamma_corrected(&self) -> Color {
@@ -129,9 +129,9 @@ impl Color {
 impl Default for Color {
     fn default() -> Color {
         Color {
-            red: 0f64,
-            green: 0f64,
-            blue: 0f64,
+            red: 0f32,
+            green: 0f32,
+            blue: 0f32,
         }
     }
 }
@@ -154,10 +154,10 @@ impl std::ops::Mul<Color> for Color {
     }
 }
 
-impl std::ops::Mul<f64> for Color {
+impl std::ops::Mul<f32> for Color {
     type Output = Color;
 
-    fn mul(self, scale: f64) -> Color {
+    fn mul(self, scale: f32) -> Color {
         Color {
             red: self.red * scale,
             green: self.green * scale,
